% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{control_boxplot_gene}
\alias{control_boxplot_gene}
\title{control_boxplot_gene}
\usage{
control_boxplot_gene(
  data,
  sel.Gene = "all",
  coef = 1.5,
  by.group = TRUE,
  colors = c("#66c2a5", "#fc8d62"),
  axis.title.size = 11,
  axis.text.size = 12,
  x.axis.title = "Gene",
  y.axis.title = "value",
  legend.title = "Group",
  legend.title.size = 11,
  legend.text.size = 11,
  legend.position = "right",
  plot.title = "",
  plot.title.size = 14,
  save.to.tiff = FALSE,
  dpi = 600,
  width = 15,
  height = 15,
  name.tiff = "control_boxplot_genes"
)
}
\arguments{
\item{data}{object returned from make_Ct_ready(), exp_Ct_dCt() or delta_Ct() functions.}

\item{sel.Gene}{character vector with names of genes to include, or "all" (default) to use all names of genes.}

\item{coef}{numeric: how many times of interquartile range should be used to indicate the most extend data point for whiskers. Default to 1.5.}

\item{by.group}{logical: if TRUE, distributions will be drawn by compared groups of samples.}

\item{colors}{character vector containing colors for groups, length of one (when by.group = FALSE) or equal to number of groups (when by.group = TRUE).}

\item{axis.title.size}{integer: font size of axis titles. Default to 12.}

\item{axis.text.size}{integer: font size of axis text. Default to 12.}

\item{x.axis.title}{character: title of x axis. Default to "Gene".}

\item{y.axis.title}{character: title of y axis. Default to "value".}

\item{legend.title}{character: title of legend. Default to "Group".}

\item{legend.title.size}{integer: font size of legend title. Default to 12.}

\item{legend.text.size}{integer: font size of legend text. Default to 12.}

\item{legend.position}{position of the legend, one of "top", "right" (default), "bottom", "left", or "none" (no legend).
See description for legend.position in ggplot2::theme() function.}

\item{plot.title}{character: title of plot. Default to "".}

\item{plot.title.size}{integer: font size of plot title. Default to 14.}

\item{save.to.tiff}{logical: if TRUE, plot will be saved as .tiff file. Default to FALSE.}

\item{dpi}{integer: resolution of saved .tiff file. Default to 600.}

\item{width}{numeric: width (in cm) of saved .tiff file. Default to 15.}

\item{height}{integer: height (in cm) of saved .tiff file. Default to 15.}

\item{name.tiff}{character: name of saved .tiff file, without ".tiff" name of extension. Default to "control_boxplot_genes".}
}
\value{
Object with boxplot illustrating distribution of data for each gene. Created plot will be also displayed on graphic device.
}
\description{
This function creates boxplot illustrating distribution of data in each gene. Could be useful to compare expression of analyzed genes.
}
\examples{
library(tidyverse)
data(data.Ct)
data.CtF <- filter_Ct(data.Ct,
                      remove.Gene = c("Gene2","Gene5","Gene6","Gene9","Gene11"),
                      remove.Sample = c("Control08","Control16","Control22"))
data.CtF.ready <- make_Ct_ready(data.CtF, imput.by.mean.within.groups = TRUE)
data.dCt <- delta_Ct(data.CtF.ready, ref = "Gene8")
control.boxplot.gene <- control_boxplot_gene(data.dCt)

}
