% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{control_boxplot_target}
\alias{control_boxplot_target}
\title{control_boxplot_target}
\usage{
control_boxplot_target(
  data,
  coef = 1.5,
  by.group = TRUE,
  colors = c("#66c2a5", "#fc8d62"),
  axis.title.size = 12,
  axis.text.size = 12,
  x.axis.title = "Target",
  y.axis.title = "value",
  legend.title = "Group",
  legend.title.size = 12,
  legend.text.size = 12,
  legend.position = "right",
  plot.title = "",
  plot.title.size = 14,
  save.to.tiff = FALSE,
  dpi = 600,
  width = 15,
  height = 15,
  name.tiff = "control_boxplot_targets"
)
}
\arguments{
\item{data}{object returned from delta_Ct, exp_Ct or exp_delta_Ct function.}

\item{coef}{numeric: how many times of interquartile range should be used to indicate the most extend data point for whiskers.}

\item{by.group}{logical: if TRUE, distributions will be drawn by compared groups of samples.}

\item{colors}{character vector length of one (when by.group = FALSE) or two (when by.group = TRUE), containing colors for groups.}

\item{axis.title.size}{integer: font size of axis titles.}

\item{axis.text.size}{integer: font size of axis text.}

\item{x.axis.title}{character: title of x axis.}

\item{y.axis.title}{character: title of y axis.}

\item{legend.title}{character: title of legend.}

\item{legend.title.size}{integer: font size of legend title.}

\item{legend.text.size}{integer: font size of legend text.}

\item{legend.position}{position of the legend, one of the following: "top", "right", "bottom", "left".}

\item{plot.title}{character: title of plot.}

\item{plot.title.size}{integer: font size of plot title.}

\item{save.to.tiff}{logical: if TRUE, plot will be saved as .tiff file.}

\item{dpi}{integer: resolution of saved .tiff file.}

\item{width}{numeric: width (in cm) of saved .tiff file.}

\item{height}{integer: height (in cm) of saved .tiff file.}

\item{name.tiff}{character: name of saved .tiff file, without ".tiff" name of extension.}
}
\value{
Boxplot illustrating distribution of data in each target.
}
\description{
Boxplot illustrating distribution of data in each target. Could be useful to compare expression of analyzed targets.
}
\examples{
control_boxplot_target(data.Ct.exp)
control_boxplot_target(data.dCt)
control_boxplot_target(data.dCt.exp)

}
