% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{control_cluster_sample}
\alias{control_cluster_sample}
\title{control_cluster_sample}
\usage{
control_cluster_sample(
  data,
  method.dist = "euclidean",
  method.clust = "complete",
  x.axis.title = "Samples",
  y.axis.title = "Height",
  plot.title = "",
  save.to.tiff = FALSE,
  dpi = 600,
  width = 15,
  height = 15,
  name.tiff = "control_clust_samples"
)
}
\arguments{
\item{data}{object returned from delta_Ct, exp_Ct or exp_delta_Ct function.}

\item{method.dist}{character: name of used method for calculation of distances, should be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski".}

\item{method.clust}{character: name of used method for agglomeration, should be one of "ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median" or "centroid".}

\item{x.axis.title}{character: title of x axis.}

\item{y.axis.title}{character: title of y axis.}

\item{plot.title}{character: title of plot.}

\item{save.to.tiff}{logical: if TRUE, plot will be saved as .tiff file.}

\item{dpi}{integer: resolution of saved .tiff file.}

\item{width}{numeric: width (in cm) of saved .tiff file.}

\item{height}{integer: height (in cm) of saved .tiff file.}

\item{name.tiff}{character: name of saved .tiff file, without ".tiff" name of extension.}
}
\value{
plot with hierarchical clustering of samples
}
\description{
Performs hierarchical clustering of samples based on the data. Could be useful to identify outlier samples.
}
\examples{
control_cluster_sample(data.dCt, method.dist = "euclidean", method.clust = "complete")
control_cluster_sample(data.dCt.exp, method.dist = "canberra", method.clust = "average")
control_cluster_sample(data.Ct.exp, method.dist = "manhattan", method.clust = "single")

}
