% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{control_corr_sample}
\alias{control_corr_sample}
\title{control_corr_sample}
\usage{
control_corr_sample(
  data,
  method = "pearson",
  add.coef = "black",
  order = "hclust",
  hclust.method = "average",
  size = 0.6,
  coef.size = 0.6,
  p.adjust.method = "BH",
  save.to.tiff = FALSE,
  dpi = 600,
  width = 15,
  height = 15,
  name.tiff = "control_corr_samples",
  save.to.txt = FALSE,
  name.txt = "control_corr_samples"
)
}
\arguments{
\item{data}{object returned from delta_Ct(), exp_Ct() or exp_delta_Ct() function.}

\item{method}{character: type of correlations to compute, specify "pearson" (default) for Pearson's correlation coefficients
or "spearman" for Spearman's rank correlation coefficients.}

\item{add.coef}{if coefficients should be add to the plot, specify color of coefficients (default to "black"), otherwise set to NULL.}

\item{order}{character: method used for ordering the correlation matrix (see documentation for corrplot::corrplot() function),
one of the "original" (original order), "AOE" (angular order of the eigenvectors),
"FPC" (first principal component order), "hclust" (hierarchical clustering order, default), or "alphabet" (alphabetical order).}

\item{hclust.method}{character: name of used method for hclust agglomeration, should be one of "ward", ward.D", "ward.D2", "single", "complete", "average" (default), "mcquitty", "median" or "centroid".}

\item{size}{numeric: size of variable names and numbers in legend. Default to 0.6.}

\item{coef.size}{numeric: size of correlation coefficients. Default to 0.6.}

\item{p.adjust.method}{character: p value correction method for multiple testing, one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY","fdr", or "none". See documentation for stats::p.adjust() function for details.}

\item{save.to.tiff}{logical: if TRUE, plot will be saved as .tiff file. Default to FALSE.}

\item{dpi}{integer: resolution of saved .tiff file. Default to 600.}

\item{width}{numeric: width (in cm) of saved .tiff file. Default to 15.}

\item{height}{integer: height (in cm) of saved .tiff file. Default to 15.}

\item{name.tiff}{character: name of saved .tiff file, without ".tiff" name of extension. Default to "control_corr_samples".}

\item{save.to.txt}{logical: if TRUE, correlation results (sorted by descending absolute values of correlation coefficients) will be saved to .txt file. Default to FALSE.}

\item{name.txt}{character: name of saved .txt file, without ".txt" name of extension.. Default to "control_corr_samples".}
}
\value{
Plot illustrating correlation matrix (displayed in graphic device) and data.frame with computed correlation coefficients and p values.
}
\description{
Performs correlation analysis of samples based on the data. Could be useful to gain insight into relationships between analyzed samples.
}
\examples{
library(tidyverse)
library(Hmisc)
library(corrplot)
data(data.Ct)
data.CtF <- filter_Ct(data.Ct,
                      remove.Target = c("Gene2","Gene5","Gene6","Gene9","Gene11"),
                      remove.Sample = c("Control08","Control16","Control22"))
data.dCt <- delta_Ct(data.CtF,
                     imput.by.mean.within.groups = TRUE,
                     ref = "Gene8")
corr.samples <- control_corr_sample(data.dCt)
head(corr.samples)

}
