% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{control_corr_target}
\alias{control_corr_target}
\title{control_corr_target}
\usage{
control_corr_target(
  data,
  method = "pearson",
  add.coef = "black",
  order = "hclust",
  hclust.method = "complete",
  size = 0.6,
  coef.size = 0.6,
  save.to.tiff = FALSE,
  dpi = 600,
  width = 15,
  height = 15,
  name.tiff = "control_corr_targets",
  save.to.txt = FALSE,
  name.txt = "control_corr_targets",
  p.adjust.method = "BH"
)
}
\arguments{
\item{data}{object returned from delta_Ct, exp_Ct or exp_delta_Ct function.}

\item{method}{character: type of correlations to compute, specify "pearson" for Pearson's correlation coefficients or "spearman" for Spearman's rank correlation coefficients.}

\item{add.coef}{if coefficients should be add to the plot, specify color of coefficients, otherwise set to NULL.}

\item{order}{character: method used for ordering the correlation matrix, one of the "original" (original order), "AOE" (angular order of the eigenvectors), "FPC" (first principal component order), "hclust" (hierarchical clustering order), or "alphabet" (alphabetical order).}

\item{hclust.method}{character: name of used method for hclust agglomeration, should be one of "ward", ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median" or "centroid".}

\item{size}{numeric: size of variable names and numbers in legend.}

\item{coef.size}{numeric: size of correlation coefficients.}

\item{save.to.tiff}{logical: if TRUE, plot will be saved as .tiff file.}

\item{dpi}{integer: resolution of saved .tiff file.}

\item{width}{numeric: width (in cm) of saved .tiff file.}

\item{height}{integer: height (in cm) of saved .tiff file.}

\item{name.tiff}{character: name of saved .tiff file, without ".tiff" name of extension.}

\item{save.to.txt}{logical: if TRUE, correlation results (sorted by descending absolute values of correlation coefficients) will be saved to .txt file.}

\item{name.txt}{character: name of saved .txt file.}

\item{p.adjust.method}{character: p value correction method for multiple testing, one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY","fdr", or "none".}
}
\value{
plot illustrating correlation matrix and data.frame with computed correlation coefficients and p values.
}
\description{
Performs correlation analysis of targets based on the data. Could be useful to gain insight into relationships between analyzed targets.
}
\examples{
control_cluster_target(data.dCt, method.dist = "euclidean", method.clust = "complete")
control_cluster_target(data.dCt.exp, method.dist = "canberra", method.clust = "average")
control_cluster_target(data.Ct.exp, method.dist = "manhattan", method.clust = "single")

}
