% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{control_Ct_barplot_target}
\alias{control_Ct_barplot_target}
\title{control_Ct_barplot_target}
\usage{
control_Ct_barplot_target(
  data,
  flag.Ct = "Undetermined",
  maxCt = 35,
  flag = "Undetermined",
  colors = c("#66c2a5", "#fc8d62"),
  x.axis.title = "",
  y.axis.title = "Number",
  axis.title.size = 12,
  axis.text.size = 10,
  legend.title = "Reliable?",
  legend.title.size = 12,
  legend.text.size = 12,
  legend.position = "top",
  plot.title = "",
  save.to.tiff = FALSE,
  dpi = 600,
  width = 15,
  height = 15,
  name.tiff = "Ct_control_barplot_for_targets"
)
}
\arguments{
\item{data}{object returned from read_Ct_long or read_Ct_wide function, or data frame containing column named "Sample" with sample names, column named "Target" with target names,
column named "Ct" with raw Ct values, column named "Group" with group names. Optionally, data frame could contain column named "Flag" with flag information (ex. "Undetermined" and "OK"),
which will be used for reliability assessment.
This parameter has no default value (must be provided).}

\item{flag.Ct}{character of a flag used for undetermined Ct values, default to "Undetermined".}

\item{maxCt}{numeric, a maximum of Ct value allowed.}

\item{flag}{character of a flag used in Flag column for values which are unreliable, default to "Undetermined".}

\item{colors}{character vector length of two, containing colors for Ct values which were labeled as reliable (first element of vector) or not (second element of vector).}

\item{x.axis.title}{character: title of x axis.}

\item{y.axis.title}{character: title of y axis.}

\item{axis.title.size}{integer: font size of axis titles.}

\item{axis.text.size}{integer: font size of axis text.}

\item{legend.title}{character: title of legend.}

\item{legend.title.size}{integer: font size of legend title.}

\item{legend.text.size}{integer: font size of legend text.}

\item{legend.position}{position of the legend, one of "top", "right", "bottom", "left".}

\item{plot.title}{character: title of plot.}

\item{save.to.tiff}{logical: if TRUE, plot will be saved as .tiff file.}

\item{dpi}{integer: resolution of saved .tiff file.}

\item{width}{numeric: width (in cm) of saved .tiff file.}

\item{height}{integer: height (in cm) of saved .tiff file.}

\item{name.tiff}{character: name of saved .tiff file.}
}
\value{
data.frame in long format ready to analysis.
}
\description{
Target-wide control of raw Ct values across groups by illustrating numbers of Ct values labeled as reliable or not by using reliability criteria (see function parameters).
This function does not perform data filtering, but only numbers Ct values labeled as reliable or not and presents them graphically.
This function Could be useful to identify targets with low number of reliable Ct values.
}
\examples{
target.Ct.control <- control_Ct_barplot_target(data.Ct)

}
