% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{control_pca_target}
\alias{control_pca_target}
\title{control_pca_target}
\usage{
control_pca_target(
  data,
  point.size = 4,
  point.shape = 19,
  alpha = 0.7,
  label.size = 3,
  hjust = 0,
  vjust = -1,
  colors = c("#66c2a5", "#fc8d62"),
  axis.title.size = 12,
  axis.text.size = 10,
  legend.text.size = 12,
  legend.title = "Group",
  legend.title.size = 12,
  legend.position = "right",
  plot.title = "",
  plot.title.size = 14,
  save.to.tiff = FALSE,
  dpi = 600,
  width = 15,
  height = 15,
  name.tiff = "control_pca_targets"
)
}
\arguments{
\item{data}{object returned from delta_Ct, exp_Ct or exp_delta_Ct function.}

\item{point.size}{numeric: size of points.}

\item{point.shape}{integer: shape of points.}

\item{alpha}{numeric: transparency of points, a value between 0 and 1.}

\item{label.size}{numeric: size of points labels.}

\item{hjust}{numeric: horizontal position of points labels,}

\item{vjust}{numeric: vertical position of points labels,}

\item{colors}{character: color of points.}

\item{axis.title.size}{integer: font size of axis titles.}

\item{axis.text.size}{integer: font size of axis text.}

\item{legend.text.size}{integer: font size of legend text.}

\item{legend.title}{character: title of legend.}

\item{legend.title.size}{integer: font size of legend title.}

\item{legend.position}{position of the legend, one of the following: "top", "right", "bottom", "left".}

\item{plot.title}{character: title of plot.}

\item{plot.title.size}{integer: font size of plot title.}

\item{save.to.tiff}{logical: if TRUE, plot will be saved as .tiff file.}

\item{dpi}{integer: resolution of saved .tiff file.}

\item{width}{numeric: width (in cm) of saved .tiff file.}

\item{height}{integer: height (in cm) of saved .tiff file.}

\item{name.tiff}{character: name of saved .tiff file, without ".tiff" name of extension.}
}
\value{
Plot illustrating spatial arrangement of samples according to coordinates of 2 components obtained from principal component analysis (PCA).
}
\description{
Performs principal component analysis (PCA) for targets and generate plot illustrating spatial arrangement of targets using 2 components. Could be useful to gain insight into similarity in expression of analyzed targets.
IMPORTANT: PCA analysis can not deal with missing values, thus all targets with at least one missing value are removed from data before analysis. It is recommended to run this function on data after imputation of missing values.
}
\examples{
control_pca_target(data.Ct.exp)
control_pca_target(data.dCt)
control_pca_target(data.dCt.exp)

}
