% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{Ct_for_control}
\alias{Ct_for_control}
\title{Ct_for_control}
\usage{
Ct_for_control(
  data,
  imput.by.mean.within.groups = TRUE,
  save.to.txt = FALSE,
  name.txt = "data.exp.Ct"
)
}
\arguments{
\item{data}{data object returned from read_Ct_long(), read_Ct_wide() or filter_Ct() function,
or data frame containing column named "Sample" with sample names, column named "Target" with target names,
column named "Ct" with raw Ct values (must be numeric), column named "Group" with group names. Any other columns could exist, but will not be used by this function.}

\item{imput.by.mean.within.groups}{logical: if TRUE, missing values will be imputed by means within groups. Default to TRUE.}

\item{save.to.txt}{logical: if TRUE, returned data will be saved to .txt file. Default to FALSE.}

\item{name.txt}{character: name of saved .txt file, without ".txt" name of extension. Default to "data.exp.Ct".}
}
\value{
Data.frame with prepared data and printed information about number and percentage of missing values.
}
\description{
This function collapses technical replicates (if present in data) by means counts and imputes missing data by means within groups (if so indicated).
These actions also prepare Ct data for control functions.
}
\examples{
library(tidyverse)
data(data.Ct)
data.CtF <- filter_Ct(data.Ct,
                      remove.Target = c("Gene2","Gene5","Gene6","Gene9","Gene11"),
                      remove.Sample = c("Control08","Control16","Control22"))
data.CtF.prepared <- Ct_for_control(data.CtF)
head(data.CtF.prepared)

}
