% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.Ct.R
\docType{data}
\name{data.Ct}
\alias{data.Ct}
\title{Gene expression dataset for RQdeltaCT package.}
\format{
A data frame with 1288 rows and 5 variables:
\describe{
  \item{\code{Sample}}{character COLUMN_DESCRIPTION}
  \item{\code{Target}}{character COLUMN_DESCRIPTION}
  \item{\code{Ct}}{character COLUMN_DESCRIPTION}
  \item{\code{Group}}{character COLUMN_DESCRIPTION}
  \item{\code{Flag}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{Dataset is available in a package directory}
}
\usage{
data.Ct
}
\description{
Dataset contains gene expression data obtained from qPCR experiments with TaqMan assays.
}
\keyword{datasets}
