% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{exp_Ct}
\alias{exp_Ct}
\title{exp_Ct}
\usage{
exp_Ct(
  data,
  imput.by.mean.within.groups = FALSE,
  save.to.txt = FALSE,
  name.txt = "data.exp.Ct"
)
}
\arguments{
\item{data}{data object returned from read_Ct_long, read_Ct_wide or filter_Ct function, or data frame containing column named "Sample" with sample names, column named "Target" with target names,
column named "Ct" with raw Ct values, column named "Group" with group names. Any other columns could exist, by will not be used by this function.}

\item{imput.by.mean.within.groups}{logical: if TRUE, missing values will be imputed by means within groups.}

\item{save.to.txt}{logical: if TRUE, returned dataset will be saved to .txt file.}

\item{name.txt}{character: name of saved .txt file.}
}
\value{
data.frame with transformed Ct values and printed information about number and percentage of missing values.
}
\description{
Collapses technical replicates (if present in data) by means, counts and imputes missing data by means within groups (if so indicated),
and finally exponentiates Ct values by using formula 2^(-Ct).
}
\examples{
data.Ct.exp.imput <- exp_Ct(data.CtF,
                            imput.by.mean.within.groups = TRUE)

}
