% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{exp_dCt}
\alias{exp_dCt}
\title{exp_delta_Ct}
\usage{
exp_dCt(data, save.to.txt = FALSE, name.txt = "data.exp.dCt")
}
\arguments{
\item{data}{data object returned from [delta_Ct()] function.}

\item{save.to.txt}{logical: if TRUE, returned data will be saved to .txt file. Default to FALSE.}

\item{name.txt}{character: name of saved .txt file, without ".txt" name of extension. Default to "data.exp.dCt".}
}
\value{
Data.frame with exponentiated dCt values.
}
\description{
This function exponentiates delta Ct (dCt) values by using formula 2^(-dCt).
}
\examples{
data.Ct.exp <- exp_Ct(data.CtF,
                      imput.by.mean.within.groups = TRUE)

}
