% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{find_ref_gene}
\alias{find_ref_gene}
\title{find_ref_gene}
\usage{
find_ref_gene(
  data,
  groups,
  candidates,
  colors,
  line.width = 1,
  angle = 0,
  x.axis.title = "",
  y.axis.title = "Ct",
  axis.title.size = 11,
  axis.text.size = 10,
  legend.title = "",
  legend.title.size = 11,
  legend.text.size = 11,
  legend.position = "top",
  plot.title = "",
  plot.title.size = 14,
  save.to.tiff = FALSE,
  dpi = 600,
  width = 15,
  height = 15,
  name.tiff = "Ct_reference_gene_selection"
)
}
\arguments{
\item{data}{object returned from make_Ct_ready() functions,}

\item{groups}{character vector length of two with names of compared groups}

\item{candidates}{character: vector of names of genes - candidates for gene reference.}

\item{colors}{character: vector of colors for genes, number of elements should be equal to number of candidate genes (elements in `candidates` vector).}

\item{line.width}{numeric: width of lines drawn in the plot. Default to 1.}

\item{angle}{integer: value of angle in which names of genes should be displayed. Default to 0.}

\item{x.axis.title}{character: title of x axis. Default to "".}

\item{y.axis.title}{character: title of y axis.  Default to "Ct".}

\item{axis.title.size}{integer: font size of axis titles. Default to 12.}

\item{axis.text.size}{integer: font size of axis text. Default to 10.}

\item{legend.title}{character: title of legend. Default to "".}

\item{legend.title.size}{integer: font size of legend title. Default to 12.}

\item{legend.text.size}{integer: font size of legend text. Default to 12.}

\item{legend.position}{position of the legend, one of "top" (default), "right", "bottom", "left", or "none" (no legend).
See description for legend.position in ggplot2::theme() function.}

\item{plot.title}{character: title of plot. Default to "".}

\item{plot.title.size}{integer: font size of plot title. Default to 14.}

\item{save.to.tiff}{logical: if TRUE, plot will be saved as .tiff file. Default to FALSE.}

\item{dpi}{integer: resolution of saved .tiff file. Default to 600.}

\item{width}{numeric: width (in cm) of saved .tiff file. Default to 15.}

\item{height}{integer: height (in cm) of saved .tiff file. Default to 15.}

\item{name.tiff}{character: name of saved .tiff file, without ".tiff" name of extension. Default to "Ct_reference_gene_selection".}
}
\value{
List containing plot object and table with calculated parameters. Created plot is displayed on graphic device.
}
\description{
This function draw a line plot and calculate parameters useful to assess gene expression stability:
 minimum, maximum, standard deviation, variance, colinearity coefficient (VIF),
 and stability measures from NormFinder and geNorm algorithms. This function could be helpful to select the best
reference gene for normalization of Ct values.
}
\examples{
library(car)
library(ctrlGene)
library(tidyverse)
data(data.Ct)
data.CtF <- filter_Ct(data.Ct,
                       remove.gene = c("Gene2","Gene5","Gene6","Gene9","Gene11"),
                       remove.Sample = c("Control08","Control16","Control22"))
data.CtF.ready <- make_Ct_ready(data.CtF, imput.by.mean.within.groups = TRUE)
ref <- find_ref_gene(data.CtF.ready, groups = c("Disease","Control"),
                          candidates = c("Gene4", "Gene8","Gene10","Gene16","Gene17", "Gene18"),
                          col = c("#66c2a5", "#fc8d62","#6A6599", "#D62728", "#1F77B4", "black"))
ref[[2]]

}
