% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{log_reg}
\alias{log_reg}
\title{log_reg}
\usage{
log_reg(
  data,
  sel.genes = "all",
  group.study,
  group.ref,
  centerline = 1,
  ci = 0.95,
  log.axis = FALSE,
  x.axis.title = "Odds ratio",
  y.axis.title = "",
  axis.title.size = 11,
  axis.text.size = 10,
  legend.title = "p value",
  legend.text.size = 11,
  legend.title.size = 11,
  legend.position = "right",
  plot.title = "",
  plot.title.size = 14,
  save.to.tiff = FALSE,
  dpi = 600,
  width = 15,
  height = 15,
  name.tiff = "OR_plot",
  save.to.txt = FALSE,
  name.txt = "OR_results"
)
}
\arguments{
\item{data}{object returned from exp_Ct_dCt() or delta_Ct() functions.}

\item{sel.genes}{character vector with names of genes to include, or "all" (default) to use all names of genes.}

\item{group.study}{character: name of study group (group of interest).}

\item{group.ref}{character: name of reference group.}

\item{centerline}{numeric: position of vertical centerline on the plot, if logaxis = TRUE, centerline should be set to 0, otherwise to 1 (default).}

\item{ci}{numeric: confidence level used for computation of confidence interval. Default to 0.95.}

\item{log.axis}{logical: if TRUE, axis with odds ratio values will be in log10 scale. If axis is in logarithmic scale, centerline parameter should be set to 0. Default ot FALSE.}

\item{x.axis.title}{character: title of x axis. Default to "gene".}

\item{y.axis.title}{character: title of y axis. Default to "value".}

\item{axis.title.size}{integer: font size of axis titles. Default to 12.}

\item{axis.text.size}{integer: font size of axis text. Default to 10.}

\item{legend.title}{character: title of legend. Default to "Group".}

\item{legend.text.size}{integer: font size of legend text. Default to 12.}

\item{legend.title.size}{integer: font size of legend title. Default to 12.}

\item{legend.position}{position of the legend, one of "top" (default), "right", "bottom", "left", or "none" (no legend).
See description for legend.position in ggplot2::theme() function.}

\item{plot.title}{character: title of plot. Default to "".}

\item{plot.title.size}{integer: font size of plot title. Default to 14.}

\item{save.to.tiff}{logical: if TRUE, plot will be saved as .tiff file. Default to FALSE.}

\item{dpi}{integer: resolution of saved .tiff file. Default to 600.}

\item{width}{numeric: width (in cm) of saved .tiff file. Default to 15.}

\item{height}{integer: height (in cm) of saved .tiff file. Default to 15.}

\item{name.tiff}{character: name of saved .tiff file, without ".tiff" name of extension. Default to "OR_plot".}

\item{save.to.txt}{logical: if TRUE, returned table with results will be saved to .txt file. Default to FALSE.}

\item{name.txt}{character: name of saved .txt file, without ".txt" name of extension. Default to "OR_results".}
}
\value{
List containing object with barplot and dataframe with results. Created plot will be also displayed on graphic device.
}
\description{
This function performs logistic regression analysis, computes odd ratio values and presents them graphically.
}
\examples{
library(tidyverse)
library(oddsratio)
data(data.Ct)
data.CtF <- filter_Ct(data.Ct,
                      remove.gene = c("Gene2","Gene5","Gene6","Gene9","Gene11"),
                      remove.Sample = c("Control08","Control16","Control22"))
data.dCt <- delta_Ct(data.CtF,
                     imput.by.mean.within.groups = TRUE,
                     ref = "Gene8")
log.reg.results <- log_reg(data.dCt,
                            sel.genes = c("Gene1","Gene16","Gene19","Gene20"),
                            group.study = "Disease",
                            group.ref = "Control")
}
