% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{low_quality_genes}
\alias{low_quality_genes}
\title{low_quality_genes}
\usage{
low_quality_genes(data, fraction = 0.5, groups)
}
\arguments{
\item{data}{object returned from control_Ct_barplot_gene() function.}

\item{fraction}{numeric: a threshold fraction (ranged from 0 to 1), genes with the higher fraction of Ct values labeled as unreliable in at least one of the compared groups will be returned. Default to 0.5.}

\item{groups}{character vector length of two, containing names of compared groups.}
}
\value{
Vector with names of genes.
}
\description{
Indicates genes with the amount of unreliable Ct values higher than specified fraction in at least one of the compared groups.
Could be useful for identification of genes with specified fraction of unreliable Ct values.
genes with high amount of unreliable data could be considered to remove from the data.
}
\examples{
library(tidyverse)
data(data.Ct)
gene.Ct.control <- control_Ct_barplot_gene(data.Ct)
low.quality.genes <- low_quality_genes(gene.Ct.control[[2]], fraction = 0.5, groups = c("Disease", "Control"))
low.quality.genes

}
