% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{make_Ct_ready}
\alias{make_Ct_ready}
\title{make_Ct_ready}
\usage{
make_Ct_ready(
  data,
  imput.by.mean.within.groups,
  save.to.txt = FALSE,
  name.txt = "Ct_ready"
)
}
\arguments{
\item{data}{data object returned from read_Ct_long(), read_Ct_wide() or filter_Ct() function,
or data frame containing column named "Sample" with sample names, column named "gene" with gene names,
column named "Ct" with raw Ct values (must be numeric), column named "Group" with group names. Any other columns could exist, but will not be used by this function.}

\item{imput.by.mean.within.groups}{logical: if TRUE, missing values will be imputed by means within groups. This parameter could influence results, thus to draw more user attention on this parameter, no default value was set.}

\item{save.to.txt}{logical: if TRUE, returned data will be saved to .txt file. Default to FALSE.}

\item{name.txt}{character: name of saved .txt file, without ".txt" name of extension. Default to "Ct_ready".}
}
\value{
Data.frame with prepared data and printed information about number and percentage of missing values.
}
\description{
This function collapses technical replicates (if present in data) by means counts and imputes missing data by means within groups (if so indicated).
These actions also prepare Ct data for control functions.
}
\examples{
library(tidyverse)
data(data.Ct)
data.CtF <- filter_Ct(data.Ct,
                      remove.gene = c("Gene2","Gene5","Gene6","Gene9","Gene11"),
                      remove.Sample = c("Control08","Control16","Control22"))
data.CtF.ready <- make_Ct_ready(data.CtF, imput.by.mean.within.groups = TRUE)
head(data.CtF.ready)

}
