% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{norm_finder}
\alias{norm_finder}
\title{norm_finder}
\usage{
norm_finder(data, candidates)
}
\arguments{
\item{data}{object returned from make_Ct_ready() functions.}

\item{candidates}{character: vector of names of targets - candidates for reference gene.}

\item{save.to.txt}{logical: if TRUE, returned table with results will be saved to .txt file. Default to FALSE.}

\item{name.txt}{character: name of saved .txt file, without ".txt" name of extension. Default to "RQ_exp_results".}
}
\value{
Table with calculated stability scores, the lowest value the best candidate for reference gene.
}
\description{
This function calculate stability scores using a code adapted from the [NormFinder algorithm](https://www.moma.dk/software/normfinder) published in [this article](https://aacrjournals.org/cancerres/article/64/15/5245/511517/Normalization-of-Real-Time-Quantitative-Reverse).
This function is internally used by `RQdeltaCT::find_ref_gene()` function and do not need to be use separately.
}
\examples{
library(tidyverse)
data(data.Ct)
data.CtF <- filter_Ct(data.Ct,
                      remove.Target = c("Gene2","Gene5","Gene6","Gene9","Gene11"),
                      remove.Sample = c("Control08","Control16","Control22"))
data.CtF.ready <- make_Ct_ready(data.CtF, imput.by.mean.within.groups = TRUE)
reference.stability.nF <- norm_finder(data.CtF.ready,
                                   candidates = c("Gene4", "Gene8","Gene10","Gene16","Gene17", "Gene18"))

}
