% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{results_barplot}
\alias{results_barplot}
\title{results_barplot}
\usage{
results_barplot(
  data,
  sel.Gene = "all",
  bar.width = 0.8,
  signif.show = TRUE,
  signif.labels,
  signif.length = 0.2,
  signif.dist = 0.2,
  faceting = TRUE,
  facet.row,
  facet.col,
  y.exp.low = 0.1,
  y.exp.up = 0.2,
  angle = 0,
  rotate = FALSE,
  colors = c("#66c2a5", "#fc8d62"),
  x.axis.title = "",
  y.axis.title = "value",
  axis.title.size = 11,
  axis.text.size = 10,
  legend.text.size = 11,
  legend.title = "Group",
  legend.title.size = 11,
  legend.position = "top",
  plot.title = "",
  plot.title.size = 14,
  save.to.tiff = FALSE,
  dpi = 600,
  width = 15,
  height = 15,
  name.tiff = "results_barplot"
)
}
\arguments{
\item{data}{object returned from make_Ct_ready(), exp_Ct_dCt() or delta_Ct() functions.}

\item{sel.Gene}{character vector with names of genes to include, or "all" (default) to use all names of genes.}

\item{bar.width}{numeric: width of bars.}

\item{signif.show}{logical: if TRUE, labels for statistical significance will be added to the plot.}

\item{signif.labels}{character vector with statistical significance labels (ex. "ns.","***", etc.) with number
of elements equal to nimber of genes used for plotting.
The ggsignif package was used for convenient adding labels, but there is one tricky point:
the same elements of labels can not be handled by used package and must be different.
It could be achieved by adding symmetrically white spaces to repeated labels, ex. "ns.", " ns. ", "  ns.  ".}

\item{signif.length}{numeric: length of horizontal bars, values from 0 to 1.}

\item{signif.dist}{numeric: distance between errorbar and significance label.
Could be in y axis units (if `faceting` = TRUE) or fraction of y axis value reached by errorbar (mean + sd value) (if `faceting` = TRUE).}

\item{faceting}{logical: if TRUE (default), plot will be drawn with facets with free scales using ggplot2::facet_wrap() function (see its documentation for more details).}

\item{facet.row, facet.col}{integer: number of rows and columns to arrange facets.}

\item{y.exp.low, y.exp.up}{numeric: space between data on the plot and lower or upper axis. Useful to add extra space for statistical significance labels when `faceting` = TRUE.}

\item{angle}{integer: value of angle in which names of genes should be displayed. Default to 0.}

\item{rotate}{logical: if TRUE, boxplots will be arranged horizontally. Deafault to FALSE.}

\item{colors}{character vector length of one (when by.group = FALSE) or two (when by.group = TRUE), containing colors for groups. Numbers of colors must be equal to number of groups. Default to c("#66c2a5", "#fc8d62").}

\item{x.axis.title}{character: title of x axis. Default to "Gene".}

\item{y.axis.title}{character: title of y axis. Default to "value".}

\item{axis.title.size}{integer: font size of axis titles. Default to 12.}

\item{axis.text.size}{integer: font size of axis text. Default to 10.}

\item{legend.text.size}{integer: font size of legend text. Default to 12.}

\item{legend.title}{character: title of legend. Default to "Group".}

\item{legend.title.size}{integer: font size of legend title. Default to 12.}

\item{legend.position}{position of the legend, one of "top" (default), "right", "bottom", "left", or "none" (no legend).
See description for legend.position in ggplot2::theme() function.}

\item{plot.title}{character: title of plot. Default to "".}

\item{plot.title.size}{integer: font size of plot title. Default to 14.}

\item{save.to.tiff}{logical: if TRUE, plot will be saved as .tiff file. Default to FALSE.}

\item{dpi}{integer: resolution of saved .tiff file. Default to 600.}

\item{width}{numeric: width (in cm) of saved .tiff file. Default to 15.}

\item{height}{integer: height (in cm) of saved .tiff file. Default to 15.}

\item{name.tiff}{character: name of saved .tiff file, without ".tiff" name of extension. Default to "results_boxplot".}
}
\value{
Object with boxplot illustrating distribution of data for selected genes. Created plot will be also displayed on graphic device.
}
\description{
This function creates a barplot illustrating mean and sd values of each gene.
Faceting and adding custom labels of statistical significance is possible.
This function could be useful to present results for finally selected genes.
}
\examples{
library(ggsignif)
library(tidyverse)
data(data.Ct)
data.CtF <- filter_Ct(data.Ct,
                      remove.Gene = c("Gene2","Gene5","Gene6","Gene9","Gene11"),
                      remove.Sample = c("Control08","Control16","Control22"))
data.CtF.ready <- make_Ct_ready(data.CtF, imput.by.mean.within.groups = TRUE)
data.dCt <- delta_Ct(data.CtF.ready, ref = "Gene8")
data.dCt.exp <- exp_delta_Ct(data.dCt)
data.dCt.expF <- filter_transformed_data(data.dCt.exp, remove.Sample = c("Control11"))
results_barplot(data.dCt.exp,
                sel.Gene = c("Gene1","Gene16","Gene19","Gene20"),
                signif.labels = c("****","*","***"," * "),
                angle = 30,
                signif.dist = 1.05,
                facet.row = 1,
                facet.col = 4,
                y.exp.up = 0.1,
                y.axis.title = bquote(~2^-dCt))

}
