% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{results_boxplot}
\alias{results_boxplot}
\title{results_boxplot}
\usage{
results_boxplot(
  data,
  coef = 1.5,
  sel.Target = "all",
  by.group = TRUE,
  faceting = TRUE,
  facet.row,
  facet.col,
  angle = 0,
  rotate = FALSE,
  add.mean = TRUE,
  add.mean.size = 2,
  add.mean.color = "black",
  colors = c("#66c2a5", "#fc8d62"),
  x.axis.title = "Target",
  y.axis.title = "value",
  axis.title.size = 12,
  axis.text.size = 10,
  legend.text.size = 12,
  legend.title = "Group",
  legend.title.size = 12,
  legend.position = "top",
  plot.title = "",
  plot.title.size = 14,
  save.to.tiff = FALSE,
  dpi = 600,
  width = 15,
  height = 15,
  name.tiff = "results_boxplot"
)
}
\arguments{
\item{data}{object returned from make_Ct_ready(), exp_Ct_or_dCt() or delta_Ct() functions.}

\item{coef}{numeric: how many times of interquartile range should be used to indicate the most extend data point for whiskers. Default to 1.5.}

\item{sel.Target}{character vector with names of targets to include, or "all" (default) to use all names of targets.}

\item{by.group}{logical: if TRUE (default), distributions will be drawn by compared groups of samples.}

\item{faceting}{logical: if TRUE (default), plot will be drawn with facets with free scales using ggplot2::facet_wrap() function (see its documentation for more details).}

\item{facet.row, facet.col}{integer: number of rows and columns to arrange facets.}

\item{angle}{integer: value of angle in which names of genes should be displayed. Default to 0.}

\item{rotate}{logical: if TRUE, boxplots will be arranged horizontally. Deafault to FALSE.}

\item{add.mean}{logical: if TRUE, means will be added to boxes as squares. Default to TRUE.}

\item{add.mean.size}{numeric: size of squares indicating means. Default to 2.}

\item{add.mean.color}{character: color of squares indicating means. Default to "black".}

\item{colors}{character vector length of one (when by.group = FALSE) or two (when by.group = TRUE), containing colors for groups.}

\item{x.axis.title}{character: title of x axis. Default to "Target".}

\item{y.axis.title}{character: title of y axis. Default to "value".}

\item{axis.title.size}{integer: font size of axis titles. Default to 12.}

\item{axis.text.size}{integer: font size of axis text. Default to 10.}

\item{legend.text.size}{integer: font size of legend text. Default to 12.}

\item{legend.title}{character: title of legend. Default to "Group".}

\item{legend.title.size}{integer: font size of legend title. Default to 12.}

\item{legend.position}{position of the legend, one of "top" (default), "right", "bottom", "left", or "none" (no legend).
See description for legend.position in ggplot2::theme() function.}

\item{plot.title}{character: title of plot. Default to "".}

\item{plot.title.size}{integer: font size of plot title. Default to 14.}

\item{save.to.tiff}{logical: if TRUE, plot will be saved as .tiff file. Default to FALSE.}

\item{dpi}{integer: resolution of saved .tiff file. Default to 600.}

\item{width}{numeric: width (in cm) of saved .tiff file. Default to 15.}

\item{height}{integer: height (in cm) of saved .tiff file. Default to 15.}

\item{name.tiff}{character: name of saved .tiff file, without ".tiff" name of extension. Default to "results_boxplot".}
}
\value{
Object with boxplot illustrating distribution of data for selected targets. Created plot will be also displayed on graphic device.
}
\description{
This function creates boxplot illustrating distribution of data in selected targets.
    It is similar to control_boxplot_target() function; however, some new options are added,
    including target selection, faceting, and adding mean labels to boxes.
    This, this function could be useful to present results for finally selected targets.
}
\examples{
library(tidyverse)
data(data.Ct)
data.CtF <- filter_Ct(data.Ct,
                      remove.Target = c("Gene2","Gene5","Gene6","Gene9","Gene11"),
                      remove.Sample = c("Control08","Control16","Control22"))
data.CtF.ready <- make_Ct_ready(data.CtF, imput.by.mean.within.groups = TRUE)
data.dCt <- delta_Ct(data.CtF.ready, ref = "Gene8")
data.dCt.exp <- exp_delta_Ct(data.dCt)
data.dCt.expF <- filter_transformed_data(data.dCt.exp, remove.Sample = c("Control11"))
results_boxplot(data.dCt.expF,
                 sel.Target = c("Gene1","Gene16","Gene19","Gene20")
                 facet.row = 2,
                 facet.col = 2,
                 y.axis.title = bquote(~2^-dCt))

}
