% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{ROCh}
\alias{ROCh}
\title{ROCh}
\usage{
ROCh(
  data,
  sel.Target = "all",
  groups,
  panels.row,
  panels.col,
  text.size = 1.1,
  print.auc = TRUE,
  print.auc.size = 0.8,
  save.to.tiff = FALSE,
  dpi = 600,
  width = 15,
  height = 15,
  name.tiff = "ROC_plot",
  save.to.txt = FALSE,
  name.txt = "ROC_results"
)
}
\arguments{
\item{data}{object returned from exp_Ct_dCt() or delta_Ct() functions.}

\item{sel.Target}{character vector with names of targets to include, or "all" (default) to use all names of targets.}

\item{groups}{character vector length of two with names of compared groups}

\item{panels.row, panels.col}{integer: number of rows and columns to arrange panels with plots.}

\item{text.size}{numeric: size of text on the plot. Default to 1.1.}

\item{print.auc}{logical: if TRUE, AUC values will be added to the plot. Default to TRUE.}

\item{print.auc.size}{numeric: size of AUC text on the plot. Default to 0.8.}

\item{save.to.tiff}{logical: if TRUE, plot will be saved as .tiff file. Default to FALSE.}

\item{dpi}{integer: resolution of saved .tiff file. Default to 600.}

\item{width}{numeric: width (in cm) of saved .tiff file. Default to 15.}

\item{height}{integer: height (in cm) of saved .tiff file. Default to 15.}

\item{name.tiff}{character: name of saved .tiff file, without ".tiff" name of extension. Default to "ROC_plot".}

\item{save.to.txt}{logical: if TRUE, returned table with results will be saved to .txt file. Default to FALSE.}

\item{name.txt}{character: name of saved .txt file, without ".txt" name of extension. Default to "ROC_results".}

\item{plot.title}{character: title of the plot.}
}
\value{
Data frame with ROC parameters including AUC, threshold, specificity, sensitivity, accuracy,
positive predictive value, negative predictive value, and Youden's J statistic.
Plot with ROC curves could be saved to .tiff file (will not be displayed on the graphic device).
}
\description{
This function performs Receiver Operating Characteristic (ROC) analysis of samples of targets based on the expression data.
This analysis could be useful to further examine performance of samples classification into groups using gene expression data.
}
\examples{
library(tidyverse)
library(pROC)
data(data.Ct)
data.CtF <- filter_Ct(data.Ct,
                      remove.Target = c("Gene2","Gene5","Gene6","Gene9","Gene11"),
                      remove.Sample = c("Control08","Control16","Control22"))
data.CtF.ready <- make_Ct_ready(data.CtF, imput.by.mean.within.groups = TRUE)
data.dCt <- delta_Ct(data.CtF.ready, ref = "Gene8")
data.dCt.expF <- filter_transformed_data(data.dCt.exp, remove.Sample = c("Control11"))
roc_parameters <- ROCh(data.dCt, sel.Target = c("Gene1","Gene16","Gene19","Gene20"),
                       groups = c("Disease","Control"),
                       panels.row = 2,
                       panels.col = 2)

}
