% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{RQ_exp_ddCt}
\alias{RQ_exp_ddCt}
\title{RQ_exp_ddCt}
\usage{
RQ_exp_ddCt(
  data,
  group.study,
  group.ref,
  do.tests = TRUE,
  save.to.txt = FALSE,
  name.txt = "ddCt_RQ_results"
)
}
\arguments{
\item{data}{data object returned from delta_Ct() function.}

\item{group.study}{character: name of study group (group of interest).}

\item{group.ref}{character: name of reference group.}

\item{do.tests}{logical: if TRUE, statistical significance of delta delta Ct values between compared groups will be calculated using Student's t test and Mann-Whitney U test. Default to TRUE.}

\item{save.to.txt}{logical: if TRUE, returned table with results will be saved to .txt file. Default to FALSE.}

\item{name.txt}{character: name of saved .txt file, without ".txt" name of extension. Default to "RQ_expCt_results".}
}
\value{
Data.frame with relative quantification results.
}
\description{
Performs relative quantification of gene expression using 2^(-ddCt) method.
}
\details{
This function calculates:
* Means (return in columns with "_mean" pattern) and standard deviations (return in columns with "_sd" pattern) of delta Ct values of analyzed targets across compared groups.
* Normality tests (Shapiro_Wilk test) of delta Ct values of analyzed targets across compared groups and returned p values in columns with "_norm_p" pattern.
* Differences in mean delta Ct values of targets between compared groups, obtaining delta delta Ct values (in returned "ddCt" column).
* RQ values (return in "RQ" column) together with log10 RQ values (return in "log10RQ" column).
  RQ values are calculated for each target by exponentiating ddCt values using 2^-ddCt formula.
* Statistical testing of delta delta Ct values (differences between study group and reference group).
  Student's t test and Mann-Whitney U test are implemented and resulted statistics (in column with "_test_stat" pattern) and p values (in column with "_test_p" pattern) are returned.
}
\examples{
library(tidyverse)
library(coin)
data(data.Ct)
data.CtF <- filter_Ct(data.Ct,
                      remove.Target = c("Gene2","Gene5","Gene6","Gene9","Gene11"),
                      remove.Sample = c("Control08","Control16","Control22"))
data.dCt <- delta_Ct(data.CtF,
                     imput.by.mean.within.groups = TRUE,
                     ref = "Gene8")
RQ.ddCt.exp <- RQ_exp_ddCt(data, "Disease", "Control")
head(RQ.ddCt.exp)

}
