% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{samples_to_remove}
\alias{samples_to_remove}
\title{samples_to_remove}
\usage{
samples_to_remove(data, fraction = 0.5)
}
\arguments{
\item{data}{object returned from [control_Ct_barplot_sample()] function.}

\item{fraction}{numeric: a threshold fraction (ranged from 0 to 1), samples with the higher fraction of Ct values labeled as unreliable will be returned. Default to 0.5.}
}
\value{
Vector with names of samples.
}
\description{
Indicates samples with the amount of unreliable Ct values higher than specified fraction.
Could be useful for identification of samples with specified fraction of unreliable Ct values.
Samples with high amount of unreliable data could be considered to remove from the dataset.
}
\examples{
samples.to.remove <- samples_to_remove(sample.Ct.control, 0.5)

}
