% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{single_pair_target}
\alias{single_pair_target}
\title{single_pair_target}
\usage{
single_pair_target(
  data,
  x,
  y,
  point.size = 4,
  point.shape = 19,
  point.alpha = 0.7,
  color = "black",
  line.alpha = 0.7,
  axis.title.size = 12,
  axis.text.size = 10,
  plot.title = "",
  plot.title.size = 14,
  labels = TRUE,
  label = c("eq", "R2", "p"),
  label.position.x = 1,
  label.position.y = 1,
  small.p = FALSE,
  small.r = FALSE,
  p.digits = 3,
  rr.digits = 2,
  save.to.tiff = FALSE,
  dpi = 600,
  width = 15,
  height = 15,
  name.tiff = "target_single_plot"
)
}
\arguments{
\item{data}{object returned from [delta_Ct()], [exp_Ct()] or [exp_delta_Ct()] function.}

\item{x, y}{characters: names of targets to use.}

\item{point.size}{numeric: size of points.}

\item{point.shape}{integer: shape of points.}

\item{point.alpha}{numeric: transparency of points, a value between 0 and 1.}

\item{color}{character: color used for points.}

\item{line.alpha}{numeric: transparency of regression line, a value between 0 and 1.}

\item{axis.title.size}{integer: font size of axis titles.}

\item{axis.text.size}{integer: font size of axis text.}

\item{plot.title}{character: title of plot.}

\item{plot.title.size}{integer: font size of plot title.}

\item{labels}{logical: if TRUE, a regression statistics will be added to the plot using [ggpimsc::stat_poly_eq()] function.}

\item{label}{character: which regression statistics should be drawn, use names specified by [ggpimsc::stat_poly_eq()] function. Default to c("eq", "R2", "p") for regression equation, coefficient of determination and p value, respectively.}

\item{label.position.x, label.position.y}{numeric: coordinates for position of regression statistics. If `by.group` = TRUE, two values could be provided for each of these parameters (for different regression lines). See description of `label.x` and `label.y` parameters from [ggpimsc::stat_poly_eq()] function.}

\item{small.p, small.r}{logical, if TRUE, p in p value label and r in coefficient of determination label, will be lowercase.}

\item{rr.digits, p.digits}{integer: number of digits after the decimal point in coefficient of determination and p value in labels.}

\item{save.to.tiff}{logical: if TRUE, plot will be saved as .tiff file.}

\item{dpi}{integer: resolution of saved .tiff file.}

\item{width}{numeric: width (in cm) of saved .tiff file.}

\item{height}{integer: height (in cm) of saved .tiff file.}

\item{name.tiff}{character: name of saved .tiff file, without ".tiff" name of extension.}
}
\value{
Plot illustrating spatial arrangement of samples according to coordinates of 2 components obtained from principal component analysis (PCA).
}
\description{
Generate scatter plot with linear regression line for two specified targets Could be useful to assess linear relationship between these targets.
}
\examples{
control_pca_target(data.Ct.exp)
control_pca_target(data.dCt)
control_pca_target(data.dCt.exp)

}
