% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{control_cluster_sample}
\alias{control_cluster_sample}
\title{control_cluster_sample}
\usage{
control_cluster_sample(
  data,
  method.dist = "euclidean",
  method.clust = "average",
  x.axis.title = "Samples",
  y.axis.title = "Height",
  plot.title = "",
  save.to.tiff = FALSE,
  dpi = 600,
  width = 15,
  height = 15,
  name.tiff = "control_clust_samples"
)
}
\arguments{
\item{data}{object returned from delta_Ct(), exp_Ct() or exp_delta_Ct() function.}

\item{method.dist}{character: name of method used for calculation of distances, derived from stats::dist() function, should be one of "euclidean" (default) , "maximum", "manhattan", "canberra", "binary" or "minkowski".}

\item{method.clust}{character: name of used method for agglomeration, derived from stats::hclust() function, should be one of "ward.D", "ward.D2", "single", "complete", "average" (default), "mcquitty", "median" or "centroid".}

\item{x.axis.title}{character: title of x axis. Default to "Samples".}

\item{y.axis.title}{character: title of y axis. Default to "Height".}

\item{plot.title}{character: title of plot. Default to "".}

\item{save.to.tiff}{logical: if TRUE, plot will be saved as .tiff file. Default to FALSE.}

\item{dpi}{integer: resolution of saved .tiff file. Default to 600.}

\item{width}{numeric: width (in cm) of saved .tiff file. Default to 15.}

\item{height}{integer: height (in cm) of saved .tiff file. Default to 15.}

\item{name.tiff}{character: name of saved .tiff file, without ".tiff" name of extension. Default to "control_clust_samples".}
}
\value{
Plot with hierarchical clustering of samples, displayed on graphic device.
}
\description{
Performs hierarchical clustering of samples based on the data. Could be useful to identify outlier samples.
}
\examples{
library(tidyverse)
data(data.Ct)
data.CtF <- filter_Ct(data.Ct,
                      remove.Target = c("Gene2","Gene5","Gene6","Gene9","Gene11"),
                      remove.Sample = c("Control08","Control16","Control22"))
data.dCt <- delta_Ct(data.CtF,
                     imput.by.mean.within.groups = TRUE,
                     ref = "Gene8")
control_cluster_sample(data.dCt)

}
