% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{control_Ct_barplot_target}
\alias{control_Ct_barplot_target}
\title{control_Ct_barplot_target}
\usage{
control_Ct_barplot_target(
  data,
  flag.Ct = "Undetermined",
  maxCt = 35,
  flag = "Undetermined",
  colors = c("#66c2a5", "#fc8d62"),
  x.axis.title = "",
  y.axis.title = "Number",
  axis.title.size = 12,
  axis.text.size = 10,
  legend.title = "Reliable Ct value?",
  legend.title.size = 12,
  legend.text.size = 12,
  legend.position = "top",
  plot.title = "",
  plot.title.size = 14,
  save.to.tiff = FALSE,
  dpi = 600,
  width = 15,
  height = 15,
  name.tiff = "Ct_control_barplot_for_targets"
)
}
\arguments{
\item{data}{object returned from [read_Ct_long()] or [read_Ct_wide()] function,
or data frame containing column named "Sample" with sample names, column named "Target" with target names,
column named "Ct" with raw Ct values, column named "Group" with group names.
Optionally, data frame could contain column named "Flag" with flag information (ex. "Undetermined" and "OK"),
which will be used for reliability assessment.}

\item{flag.Ct}{character of a flag used for undetermined Ct values. Default to "Undetermined".}

\item{maxCt}{numeric, a maximum of Ct value allowed. Default to 35.}

\item{flag}{character of a flag used in Flag column for values which are unreliable. Default to "Undetermined".}

\item{colors}{character vector length of two, containing colors for Ct values which were labeled as reliable (first element of vector) or not (second element of vector).}

\item{x.axis.title}{character: title of x axis. Default to "".}

\item{y.axis.title}{character: title of y axis. Default to "Number".}

\item{axis.title.size}{integer: font size of axis titles. Default to 12.}

\item{axis.text.size}{integer: font size of axis text. Default to 10.}

\item{legend.title}{character: title of legend. Default to "Reliable Ct value?".}

\item{legend.title.size}{integer: font size of legend title.  Default to 12.}

\item{legend.text.size}{integer: font size of legend text. Default to 12.}

\item{legend.position}{position of the legend, one of "top" (default), "right", "bottom", "left", or "none" (no legend).
See description for `legend.position` in [ggplot2::theme()] function.}

\item{plot.title}{character: title of plot. Default to "".}

\item{plot.title.size}{integer: font size of plot title. Default to 14.}

\item{save.to.tiff}{logical: if TRUE, plot will be saved as .tiff file.  Default to FALSE.}

\item{dpi}{integer: resolution of saved .tiff file. Default to 600.}

\item{width}{numeric: width (in cm) of saved .tiff file. Default to 15.}

\item{height}{integer: height (in cm) of saved .tiff file. Default to 15.}

\item{name.tiff}{character: name of saved .tiff file, without ".tiff" name of extension.  Default to "Ct_control_barplot_for_targets".}
}
\value{
list with plot and table containing numbers of reliable and not reliable Ct values in targets.
Additional information about returned table is also printed, it could help user to properly interpret returned table.
Plot will be displayed on graphic device.
}
\description{
Target-wide control of raw Ct values across groups by illustrating numbers of Ct values labeled as reliable or not by using reliability criteria (see function parameters).
}
\details{
This function does not perform data filtering, but only numbers Ct values labeled as reliable or not and presents them graphically.
Could be useful to identify targets with low number of reliable Ct values.
}
\examples{
target.Ct.control <- control_Ct_barplot_target(data.Ct)

}
