% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{exp_dCt}
\alias{exp_dCt}
\title{exp_delta_Ct}
\usage{
exp_dCt(data, save.to.txt = FALSE, name.txt = "data.exp.dCt")
}
\arguments{
\item{data}{data object returned from delta_Ct() function.}

\item{save.to.txt}{logical: if TRUE, returned data will be saved to .txt file. Default to FALSE.}

\item{name.txt}{character: name of saved .txt file, without ".txt" name of extension. Default to "data.exp.dCt".}
}
\value{
Data.frame with exponentiated dCt values.
}
\description{
This function exponentiates delta Ct (dCt) values by using formula 2^(-dCt).
}
\examples{
library(tidyverse)
data(data.Ct)
data.CtF <- filter_Ct(data.Ct,
                      remove.Target = c("Gene2","Gene5","Gene6","Gene9","Gene11"),
                      remove.Sample = c("Control08","Control16","Control22"))
data.dCt <- delta_Ct(data.CtF,
                     imput.by.mean.within.groups = TRUE,
                     ref = "Gene8")
data.dCt.exp <- exp_delta_Ct(data.dCt)
head(data.dCt.exp)

}
