% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{filter_Ct}
\alias{filter_Ct}
\title{filter_Ct}
\usage{
filter_Ct(
  data,
  flag.Ct = "Undetermined",
  maxCt = 35,
  flag = c("Undetermined"),
  remove.Target = c(""),
  remove.Sample = c(""),
  remove.Group = c("")
)
}
\arguments{
\item{data}{object returned from [read_Ct_long()] or [read_Ct_wide()] function,
or data frame containing column named "Sample" with sample names, column named "Target" with target names,
column named "Ct" with raw Ct values, column named "Group" with group names.
Optionally, data frame could contain column named "Flag" with flag information (ex. "Undetermined" and "OK"),
which will be used for filtering.}

\item{flag.Ct}{character of a flag used for undetermined Ct values, default to "Undetermined".}

\item{maxCt}{numeric, a maximum of Ct value allowed.}

\item{flag}{character: flag used in Flag column for values which should be filtered out, default to "Undetermined".}

\item{remove.Target}{character: vector with names of targets which should be removed from data}

\item{remove.Sample}{character: vector with names of samples which should be removed from data}

\item{remove.Group}{character: vector with names of groups which should be removed from data}
}
\value{
Data.frame with filtered data.
}
\description{
Filters Ct data according to the used filtering criteria (see parameters).
}
\examples{

data.CtF <- filter_Ct(data.Ct,
                      flag.Ct = "Undetermined",
                      maxCt = 35,
                      flag = "Undetermined",
                      remove.Target = c("Gene2","Gene6"),
                      remove.Sample = c("Control8","control16"))

data.CtF <- filter_Ct(data.Ct,
                       flag.Ct = "Undetermined",
                       maxCt = 35,
                       flag = "Undetermined",
                       remove.Target = targets.to.remove,
                       remove.Sample = samples.to.remove)

}
