% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{filter_transformed_Ct}
\alias{filter_transformed_Ct}
\title{filter_transformed_Ct}
\usage{
filter_transformed_Ct(
  data,
  remove.Target = c(""),
  remove.Sample = c(""),
  remove.Group = c("")
)
}
\arguments{
\item{data}{object returned from delta_Ct(), exp_Ct() or exp_delta_Ct() functions.}

\item{remove.Target}{character: vector with names of targets which should be removed from data.}

\item{remove.Sample}{character: vector with names of samples which should be removed from data.}

\item{remove.Group}{character: vector with names of groups which should be removed from data.}
}
\value{
Data.frame with filtered data.
}
\description{
Filters transformed Ct data (2^(-Ct), delta Ct, and 2^(-dCt) data) according to the used filtering criteria (see parameters).
}
\examples{
library(tidyverse)
data(data.Ct)
data.CtF <- filter_Ct(data.Ct,
                      remove.Target = c("Gene2","Gene5","Gene6","Gene9","Gene11"),
                      remove.Sample = c("Control08","Control16","Control22"))
data.dCt <- delta_Ct(data.CtF,
                     imput.by.mean.within.groups = TRUE,
                     ref = "Gene8")
data.dCt.exp <- exp_delta_Ct(data.dCt)
data.dCt.expF <- filter_transformed_Ct(data.dCt.exp, remove.Sample = c("Control11"))0

dim(data.dCt.exp)
dim(data.dCt.expF)


}
