% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQdeltaCT.R
\name{read_Ct_long}
\alias{read_Ct_long}
\title{Title}
\usage{
read_Ct_long(
  path,
  sep,
  dec,
  skip = 0,
  col.Sample,
  col.Target,
  col.Ct,
  col.Group,
  add.col.Flag = FALSE,
  col.Flag
)
}
\arguments{
\item{col.Flag}{}
}
\description{
Title
}
